requirements.txt
requests
python-dotenv
pynput